@@ -13,17 +13,17 @@
     "test": "ng test"
   },
   "dependencies": {
-    "@angular/animations": "^8.2.4",
-    "@angular/common": "^8.2.4",
-    "@angular/compiler": "^8.2.4",
-    "@angular/core": "^8.2.4",
-    "@angular/forms": "^8.2.4",
-    "@angular/platform-browser": "^8.2.4",
-    "@angular/platform-browser-dynamic": "^8.2.4",
-    "@angular/router": "^8.2.4",
+    "@angular/animations": "^9.1.4",
+    "@angular/common": "^9.1.4",
+    "@angular/compiler": "^9.1.4",
+    "@angular/core": "^9.1.4",
+    "@angular/forms": "^9.1.4",
+    "@angular/platform-browser": "^9.1.4",
+    "@angular/platform-browser-dynamic": "^9.1.4",
+    "@angular/router": "^9.1.4",
     "@aspnet/signalr": "^1.1.4",
     "@fortawesome/fontawesome-free": "^5.13.0",
-    "abp-ng2-module": "^5.1.1",
+    "abp-ng2-module": "^6.0.0",
     "abp-web-resources": "^5.1.1",
     "admin-lte-css-only": "^3.0.4",
     "block-ui": "^2.70.1",
@@ -31,45 +31,46 @@
     "famfamfam-flags": "^1.0.0",
     "jquery": "^3.4.1",
     "lodash": "^4.17.4",
-    "moment": "^2.18.1",
-    "moment-timezone": "^0.5.13",
+    "moment": "2.24.0",
+    "moment-timezone": "^0.5.28",
     "ngx-bootstrap": "^5.6.1",
-    "ngx-pagination": "^4.1.0",
-    "push.js": "1.0.9",
+    "ngx-pagination": "^5.0.0",
+    "push.js": "1.0.12",
     "rxjs": "^6.4.0",
     "spin.js": "^2.3.2",
     "sweetalert": "^2.0.8",
     "toastr": "^2.1.2",
     "ts-helpers": "^1.1.2",
+    "tslib": "^1.10.0",
     "web-animations-js": "^2.3.2",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.2",
-    "@angular/cli": "^8.3.2",
-    "@angular/compiler-cli": "^8.2.4",
+    "@angular-devkit/build-angular": "~0.901.4",
+    "@angular/cli": "^9.1.4",
+    "@angular/compiler-cli": "^9.1.4",
     "@angularclass/hmr": "^2.1.3",
     "@types/jasmine": "^3.4.0",
-    "@types/jquery": "^3.3.34",
+    "@types/jquery": "^3.3.37",
     "@types/jquery.blockui": "0.0.28",
-    "@types/lodash": "^4.14.62",
+    "@types/lodash": "^4.14.150",
     "@types/moment-timezone": "^0.5.12",
-    "@types/node": "^12.7.3",
-    "codelyzer": "^5.0.1",
+    "@types/node": "^13.13.4",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "^3.3.0",
-    "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^4.3.0",
+    "jasmine-spec-reporter": "^5.0.2",
+    "karma": "^5.0.4",
     "karma-chrome-launcher": "^3.1.0",
     "karma-cli": "^2.0.0",
-    "karma-coverage-istanbul-reporter": "^2.0.4",
-    "karma-jasmine": "^2.0.1",
+    "karma-coverage-istanbul-reporter": "^3.0.0",
+    "karma-jasmine": "^3.1.1",
     "karma-jasmine-html-reporter": "^1.4.0",
-    "nswag": "^13.0.6",
-    "protractor": "^5.1.1",
-    "ts-node": "^8.3.0",
-    "tslint": "^5.19.0",
-    "typescript": "3.4.5",
-    "webpack-bundle-analyzer": "^3.0.3"
+    "nswag": "^13.4.2",
+    "protractor": "^5.4.4",
+    "ts-node": "^8.10.1",
+    "tslint": "^6.1.2",
+    "typescript": "3.8.3",
+    "webpack-bundle-analyzer": "^3.7.0"
   },
   "angular-cli": {}
 }
